pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "plastiglom"
version = "0.0.1"
description = "A structured daily self-reflection system with LLM-assisted analysis."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Plastiglom" }]
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "anthropic>=0.40",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]
telegram = ["python-telegram-bot>=21"]
web = ["fastapi>=0.110", "uvicorn>=0.29", "jinja2>=3.1"]

[project.scripts]
plastiglom-scheduler = "plastiglom.apps.scheduler.__main__:main"
plastiglom-archiver = "plastiglom.apps.archiver.__main__:main"
plastiglom-tagger = "plastiglom.apps.tagger.__main__:main"
plastiglom-analyzer = "plastiglom.apps.analyzer.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/plastiglom"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
